Introduce Github workflow
This commit introduces a new workflow for f.el that runs its tests on
each new commit as well as on pull requests.

It seems that support for Emacs 24 needs to be dropped, Cask refuses
to install the required dependencies for any Emacs version below 25.1.
While they all should be compatible with at least Emacs 24.5,
`package-build' requires Emacs 25.1 and is a dependency of Cask itself.